+++ src/components/Dialog/index.ts
@@ -1,2 +1,3 @@
 export * from './Dialog';
+export * from './DialogHelper';
 export * from './Dialog.types';